moved confirm_bulk_delete into journalist_app/utils.py
